+++ R/badge.R
@@ -44,49 +44,19 @@ NULL
 #' @rdname badges
 #' @export
 use_badge <- function(badge_name, href, src) {
-  link <- glue("[![{badge_name}]({src})]({href})")
-  badge_lines <- badge_lines()
-
-  if (is.null(badge_lines)) {
-    if (has_badge(href)) {
-      return(invisible(FALSE))
-    }
-
-    todo(
-      "Add a {field(badge_name)} badge by adding the following line ",
-      "to your README:"
-    )
-    code_block(link)
-    todo(
-      "Allow badges to be added automatically by adding ",
-      "{code(badge_start)} and {code(badge_end)} before and after your badges."
-    )
-    return(invisible(FALSE))
-  }
-
   path <- find_readme()
-  lines <- readLines(path)
-  start <- badge_lines[[1]]
-  end <- badge_lines[[2]]
-
-  badges <- lines[seq2(start, end)]
-  if (link %in% badges) {
-    return(invisible(FALSE))
-  }
-
-  done("Adding a {field(badge_name)} badge to {value(proj_rel_path(path))}")
-  badges <- c(badges, link)
-  lines <- c(
-    lines[seq2(1, start - 1L)],
-    badges,
-    lines[seq2(end + 1L, length(lines))]
+  changed <- block_append(
+    glue("{field(badge_name)} badge"),
+    glue("[![{badge_name}]({src})]({href})"),
+    path = path,
+    block_start = badge_start,
+    block_end = badge_end
   )
-  writeLines(lines, path)
 
-  if (path_ext(path) == "Rmd") {
+  if (changed && path_ext(path) == "Rmd") {
     todo("Re-knit {value(proj_rel_path(path))}")
   }
-  invisible(TRUE)
+  invisible(changed)
 }
 
 #' @rdname badges
@@ -168,29 +138,6 @@ has_badge <- function(href) {
 
 badge_start <- "<!-- badges: start -->"
 badge_end <- "<!-- badges: end -->"
-
-badge_lines <- function() {
-  path <- find_readme()
-  if (is.null(path))
-    return(NULL)
-
-  lines <- readLines(path)
-  start <- which(lines == badge_start)
-  end <- which(lines == badge_end)
-
-  # No badges
-  if (length(start) == 0 && length(end) == 0)
-    return(NULL)
-
-  if (!(length(start) == 1 && length(end) == 1 && start < end)) {
-    stop_glue(
-      "Invalid badge specification.
-      Must start with {code(badge_start)} and end with {code(badge_end)}"
-    )
-  }
-
-  c(start + 1L, end - 1L)
-}
 
 find_readme <- function() {
   Rmd <- proj_path("README.Rmd")